Create oil
This code uses the axios library to make a request to the Alpha Vantage API to retrieve the latest daily adjusted time series data for the "OIL" symbol. It then accesses the "4. close" key of the latest data point, which represents the latest closing price of oil. It logs the latest price to the console.
You will need to get your own API key from alpha vantage to use this snippet